requirements.txt
# Django and DRF
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
django-cors-headers>=4.1.0,<5.0.0
django-filter>=23.2,<24.0
django-taggit>=4.0.0,<5.0.0
django-ipware>=5.0.0,<6.0.0
django-admin-interface>=0.30.0,<0.31.0

# Swagger/API documentation
drf-yasg>=1.21.7,<2.0.0

# Image processing
Pillow>=10.0.0,<11.0.0

# Development & Testing
pytest>=7.4.0,<8.0.0
pytest-django>=4.5.2,<5.0.0
coverage>=7.3.0,<8.0.0
flake8>=6.1.0,<7.0.0
black>=23.7.0,<24.0.0
factory-boy>=3.3.0,<4.0.0  # For test factories
faker>=19.0.0,<20.0.0  # For fake data generation

# Production
gunicorn>=21.2.0,<22.0.0
psycopg2-binary>=2.9.7,<3.0.0  # PostgreSQL adapter
whitenoise>=6.5.0,<7.0.0  # Static files serving
redis>=5.0.0,<6.0.0  # For caching and message broker
django-redis>=5.3.0,<6.0.0  # Redis cache backend
dj-database-url>=2.1.0,<3.0.0  # Database URL configuration

# Performance & Monitoring
psutil>=5.9.0,<6.0.0  # System monitoring
django-debug-toolbar>=4.2.0,<5.0.0  # Development debugging

# Security
django-ratelimit>=4.1.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0

# Search (PostgreSQL full-text search)
# Note: Uses built-in PostgreSQL features, no additional packages needed 